+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Changed
+
+- Handle dynamic imports that use dereferencing
+
 ## [6.0.0] - 24 Jan 2020
 
 ### Added