+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changed
 
+- Fix to support destructuerd exports like 'export const { a, b }'
 - Handle dynamic imports that use dereferencing
 
 ## [6.0.0] - 24 Jan 2020